\name{fenceline.ex}
\alias{fenceline.ex}
\docType{data}
\title{
Example specification of a site fenceline or boundary
}
\description{
A list(x,y) suitable as an argument to the polygon() function.
}
\usage{data(fenceline.ex)}
\format{
  The format is:
List of 2
 $ x: num [1:8] 2097849 2098180 2099364 2101096 2101287 ...
 $ y: num [1:8] 859218 861423 861765 861361 859777 ...
}
\details{
This object can be used to add a line representing a site boundary to
other plots.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(fenceline.ex)
}
\keyword{datasets}
